+++ src/Captura.Core/Models/CommandLine/CmdOptionsBase.cs
@@ -1,0 +1,17 @@
+﻿using CommandLine;
+using CommandLine.Text;
+
+namespace Captura
+{
+    public class CmdOptionsBase
+    {
+        [ParserState]
+        public IParserState LastParserState { get; set; }
+
+        [HelpOption]
+        public string GetUsage()
+        {
+            return HelpText.AutoBuild(this, current => HelpText.DefaultParsingErrorsHandler(this, current));
+        }
+    }
+}
