+++ src/Captura.Core/Models/CommandLine/CommonCmdOptions.cs
@@ -1,0 +1,19 @@
+﻿using CommandLine;
+
+namespace Captura
+{
+    public class CommonCmdOptions : CmdOptionsBase
+    {
+        [Option("reset", DefaultValue = false, HelpText = "Reset all setting values to default.")]
+        public bool Reset { get; set; }
+
+        [Option("tray", DefaultValue = false, HelpText = "Start minimized into the system tray.")]
+        public bool Tray { get; set; }
+
+        [Option("delay", DefaultValue = 0, HelpText = "Milliseconds to wait before starting recording.")]
+        public int Delay { get; set; }
+
+        [Option("length", DefaultValue = 0, HelpText = "Length of Recording in seconds.")]
+        public int Length { get; set; }        
+    }
+}
